refactor: move all filter methods to Filter class
